backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
google-generativeai==0.3.2
googlemaps==4.10.0
aiohttp==3.9.3
pydantic==2.6.1
python-multipart==0.0.6
requests==2.31.0